Refactor ssytem
